requirements.txt
flask==2.3.3
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
groq==0.4.0
gunicorn==21.2.0
python-dotenv==1.0.0 